changed how we give resume and job description  to gemini. Now we give it the raw pdf.
Also instead of scraping job application form and programmatically deciding how to map resume values to form fields, we give the raw pdf resume and screenshots of the application form to gemini for mapping.

Also added form to the bottom of dashboard for common questions not found on resume but found on job applications.

getting closer to getting all form fields to be auto filled